// Katiba55.Client/src/app/enums/officer-rank.enum.ts
export enum OfficerRank {
    SecondLieutenant = 'SecondLieutenant', // ملازم
    FirstLieutenant = 'FirstLieutenant', // ملازم اول
    Captain = 'Captain', // نقيب
    Major = 'Major', // رائد
    LieutenantColonel = 'LieutenantColonel', // مقدم
    Colonel = 'Colonel', // عقيد
    Brigadier = 'Brigadier', // عميد
    MajorGeneral = 'MajorGeneral', // لواء
    LieutenantGeneral = 'LieutenantGeneral', // فريق
    General = 'General', // فريق اول
    FieldMarshal = 'FieldMarshal' // مشير
}